removed default page and updated login page
