--- nba-front-end/src/components/Default.tsx
@@ -1,20 +1,0 @@
-import { Button } from '@mui/material';
-import * as React from 'react';
-import { Link } from 'react-router-dom';
-
-const linkStyle = {
-  textDecoration: "none",
-  color: 'white'
-};
-
-function DefaultPage() {
-  return (
-    <div>
-      <Button variant="contained" sx={{'margin':2}}>
-        <Link style={linkStyle} to="/dashboard/Players">Login</Link>
-      </Button>
-    </div>
-  );
-}
-
-export default DefaultPage;
