+++ nba-front-end/src/routes/routes.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
 import BasePage from "../components/Base";
-import DefaultPage from "../components/Default";
 import LoginPage from "../components/LoginPage/LoginPage";
 import Dashboard from "../components/Menu";
 import PlayerPage from '../components/PlayerPage/PlayerPage';
@@ -12,8 +11,7 @@ function RouteConfig() {
   return (
     <Routes>
         <Route path="/" element={<BasePage />}>
-            <Route path="/" element={<DefaultPage />} />
-            <Route path="login" element={<LoginPage />} />
+            <Route path="/" element={<LoginPage />} />            
             <Route path="dashboard" element={<Dashboard />}>
                 <Route path="players" element={<PlayerPage />} />
                 <Route path="teams" element={<TeamPage />} />
@@ -22,7 +20,7 @@ function RouteConfig() {
             {/* Using path="*"" means "match anything", so this route
                 acts like a catch-all for URLs that we don't have explicit
                 routes for. */}
-            <Route path="*" element={<DefaultPage />} />
+            <Route path="*" element={<LoginPage />} />
         </Route>
     </Routes>
   );
